Move Dashboard code to src/workspaces/dashboard.
