+++ src/libs/routes.js
@@ -27,13 +27,13 @@ import * as UploadData from 'src/pages/UploadDataPage';
 import * as WorkflowsList from 'src/pages/workflows/List';
 import * as WorkflowDetails from 'src/pages/workflows/workflow/WorkflowDetails';
 import * as WorkspaceList from 'src/pages/workspaces/List';
-import * as Dashboard from 'src/pages/workspaces/workspace/Dashboard/Dashboard';
 import * as WorkspaceFiles from 'src/pages/workspaces/workspace/Files';
 import * as JobHistory from 'src/pages/workspaces/workspace/JobHistory';
 import * as SubmissionDetails from 'src/pages/workspaces/workspace/jobHistory/SubmissionDetails';
 import * as WorkflowDashboard from 'src/pages/workspaces/workspace/jobHistory/WorkflowDashboard';
 import * as Workflows from 'src/pages/workspaces/workspace/Workflows';
 import * as WorkflowView from 'src/pages/workspaces/workspace/workflows/WorkflowView';
+import * as Dashboard from 'src/pages/workspaces/WorkspaceDashboard';
 import * as WorkspaceMigration from 'src/pages/workspaces/WorkspaceMigration';
 import * as PrivacyPolicy from 'src/registration/privacy-policy/PrivacyPolicy';
 import * as TermsOfService from 'src/registration/terms-of-service/TermsOfServicePage';
