+++ src/workspaces/dashboard/DashboardAuthContainer.tsx
@@ -6,8 +6,8 @@ import { Ajax } from 'src/libs/ajax';
 import { useStore } from 'src/libs/react-utils';
 import { authStore } from 'src/libs/state';
 import SignIn from 'src/pages/SignIn';
-import { WorkspaceDashboardPage } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceDashboardPage';
-import DashboardPublic from 'src/pages/workspaces/workspace/DashboardPublic';
+import DashboardPublic from 'src/workspaces/dashboard/DashboardPublic';
+import { WorkspaceDashboardPage } from 'src/workspaces/dashboard/WorkspaceDashboardPage';
 
 export interface DashboardAuthContainerProps {
   namespace: string;
