+++ src/workspaces/dashboard/WorkspaceDashboard.test.tsx
@@ -1,9 +1,9 @@
 import { screen } from '@testing-library/react';
 import React from 'react';
-import { WorkspaceDashboard } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceDashboard';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultGoogleBucketOptions, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
 import { InitializedWorkspaceWrapper as Workspace, StorageDetails } from 'src/workspaces/common/state/useWorkspace';
+import { WorkspaceDashboard } from 'src/workspaces/dashboard/WorkspaceDashboard';
 
 describe('WorkspaceDashboard', () => {
   const storageDetails: StorageDetails = {
